Tease out github_services_test.py
